Refactor submissionss and its nested end points to use the new routing style
