+++ app/controllers/api/v8/apidocs_controller.rb
@@ -96,12 +96,16 @@ class Api::V8::ApidocsController < ActionController::Base
     Api::V8::Courses::ExercisesController,
     Api::V8::Courses::Exercises::PointsController,
     Api::V8::Courses::Exercises::Users::PointsController,
+    Api::V8::Courses::SubmissionsController,
+    Api::V8::Courses::Users::SubmissionsController,
     Api::V8::Organizations::CoursesController,
     Api::V8::Organizations::Courses::PointsController,
     Api::V8::Organizations::Courses::Users::PointsController,
     Api::V8::Organizations::Courses::ExercisesController,
     Api::V8::Organizations::Courses::Exercises::PointsController,
     Api::V8::Organizations::Courses::Exercises::Users::PointsController,
+    Api::V8::Organizations::Courses::SubmissionsController,
+    Api::V8::Organizations::Courses::Users::SubmissionsController,
     Api::V8::ExercisesController,
     Api::V8::SubmissionsController,
     Api::V8::UsersController,
