+++ config/routes.rb
@@ -50,6 +50,7 @@ TmcServer::Application.routes.draw do
           resources :users, module: :courses, only: [] do
             resources :points, module: :users, only: :index
           end
+
           resources :exercises, module: :courses, param: :name, only: :index do
             resources :points, module: :exercises, only: :index
             resources :users, module: :exercises, only: [] do
@@ -57,49 +58,35 @@ TmcServer::Application.routes.draw do
             end
             get 'download', on: :member
           end
+
+          resources :submissions, module: :courses, only: :index
+          resources :users, module: :courses, only: [] do
+            resources :submissions, module: :users, only: :index
+          end
         end
       end
+
       resources :courses, only: :show do
         resources :points, module: :courses, only: :index
         resources :users, module: :courses, only: [] do
           resources :points, module: :users, only: :index
         end
+
         resources :exercises, module: :courses, param: :name, only: :index do
           resources :points, module: :exercises, only: :index
           resources :users, module: :exercises, only: [] do
             resources :points, module: :users, only: :index
           end
         end
-      end
 
-      scope '/org' do
-        scope '/:slug' do
-          scope '/courses' do
-            scope '/:course_name' do
-              scope '/submissions' do
-                get '/' => 'submissions#get_submissions_all'
-                scope '/user' do
-                  get '/' => 'submissions#get_submissions_user'
-                  get '/:user_id' => 'submissions#get_submissions_user'
-                end
-              end
-            end
-          end
+        resources :submissions, module: :courses, only: :index
+        resources :users, module: :courses, only: [] do
+          resources :submissions, module: :users, only: :index
         end
       end
-      scope '/courses' do
-        scope '/:course_id' do
-            scope '/submissions' do
-              get '/' => 'submissions#get_submissions_all'
-              scope '/user' do
-                get '/' => 'submissions#get_submissions_user'
-                get '/:user_id' => 'submissions#get_submissions_user'
-              end
-            end
-          end
-        end
     end
   end
+
   resources :organizations, except: [:destroy, :create, :edit, :update], path: 'org' do
 
     resources :exercises, only: [:show] do
